int32 -> uint8 + try decreasing resources on sell

+++ Source/Honeycomber/Private/Actors/Interactables/Desk.cpp
@@ -21,7 +21,7 @@ void ADesk::ClearDeskOptions()
 	InteractionPopup->ClearOptions();
 }
 
-void ADesk::InteractOption(int32 index)
+void ADesk::InteractOption(uint8 index)
 {
 	OnVisitorResponseDelegate.ExecuteIfBound(index);
 }
